chatbot/requirements.txt
fastapi
uvicorn
groq
python-dotenv
prometheus-fastapi-instrumentator
pydantic